Put zero sensitive float array back in
