+++ test/test_misc.jl
@@ -1,7 +1,7 @@
 using Base.Test
 
 import Celeste.Model: patch_ctrs_pix, patch_radii_pix
-
+import Celeste.SensitiveFloats.zero_sensitive_float_array
 
 function test_tile_image()
     blob, ea, three_bodies = gen_three_body_dataset()
@@ -182,6 +182,31 @@ function test_sky_noise_estimates()
 end
 
 
+function test_zero_sensitive_float_array()
+    S = 3
+    sf_vec = zero_sensitive_float_array(CanonicalParams, Float64, S, 3, 5)
+    @test size(sf_vec) == (3, 5)
+    for sf in sf_vec
+        @test sf.v[1] == 0
+        @test all([ all(sf_d .== 0) for sf_d in sf.d ])
+        @test all(sf.h .== 0)
+    end
+
+    # Test that the sensitive floats are distinct and not pointers to the
+    # same sensitive float.
+    sf_vec[1].v[1] = rand()
+    sf_vec[1].d[:, 1] = rand(length(CanonicalParams))
+    sf_vec[1].h = 2 * diagm(ones(size(sf_vec[1].h, 1)))
+
+    for ind in 2:length(sf_vec)
+        sf = sf_vec[ind]
+        @test sf.v[1] == 0
+        @test all([ all(sf_d .== 0) for sf_d in sf.d ])
+        @test all(sf.h .== 0)
+    end
+end
+
+
 ####################################################
 
 test_tile_image()
@@ -189,3 +214,4 @@ test_sky_noise_estimates()
 test_local_sources()
 test_local_sources_2()
 test_local_sources_3()
+test_zero_sensitive_float_array()
